+++ server/impl/src/main/java/com/walmartlabs/concord/server/user/UserResourceImpl.java
@@ -1,13 +1,16 @@
 package com.walmartlabs.concord.server.user;
 
-import com.walmartlabs.concord.server.api.user.*;
 import com.walmartlabs.concord.server.api.security.Permissions;
+import com.walmartlabs.concord.server.api.user.*;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.sonatype.siesta.Resource;
 import org.sonatype.siesta.Validate;
+import org.sonatype.siesta.ValidationErrorsException;
 
 import javax.inject.Inject;
 import javax.inject.Named;
+import javax.ws.rs.WebApplicationException;
+import javax.ws.rs.core.Response.Status;
 import java.util.UUID;
 
 @Named
@@ -24,6 +27,10 @@ public class UserResourceImpl implements UserResource, Resource {
     @Validate
     @RequiresPermissions(Permissions.USER_CREATE_NEW)
     public CreateUserResponse create(CreateUserRequest request) {
+        if (userDao.exists(request.getUsername())) {
+            throw new ValidationErrorsException("The user already exists: " + request.getUsername());
+        }
+
         String id = UUID.randomUUID().toString();
         userDao.insert(id, request.getUsername(), request.getPermissions());
         // TODO password?
@@ -33,6 +40,10 @@ public class UserResourceImpl implements UserResource, Resource {
     @Override
     @RequiresPermissions(Permissions.USER_DELETE_ANY)
     public DeleteUserResponse delete(String id) {
+        if (!userDao.existsById(id)) {
+            throw new WebApplicationException("User not found: " + id, Status.NOT_FOUND);
+        }
+
         userDao.delete(id);
         return new DeleteUserResponse();
     }
@@ -40,6 +51,10 @@ public class UserResourceImpl implements UserResource, Resource {
     @Override
     @RequiresPermissions(Permissions.USER_UPDATE_ANY)
     public UpdateUserResponse update(String id, UpdateUserRequest request) {
+        if (!userDao.existsById(id)) {
+            throw new WebApplicationException("User not found: " + id, Status.NOT_FOUND);
+        }
+
         userDao.update(id, request.getPermissions());
         return new UpdateUserResponse();
     }
